more 32 to 16 bit code - most of the sound code, which is far-reaching... also introduced some typedefs for things that will remain 32 bit, including converting more things to angle_t and fixed_t explicitly

+++ d_main.c
@@ -549,8 +549,8 @@ void D_DoomLoop (void)
 //
 //  DEMO LOOP
 //
-int32_t             demosequence;
-int32_t             pagetic;
+int8_t             demosequence;
+int16_t             pagetic;
 int8_t                    *pagename;
